requirements.txt
click>=8.1.0
python-dotenv>=1.0.0
cryptography>=41.0.0
validators>=0.22.0


